+++ src/utils.cpp
@@ -1,8 +1,8 @@
 #include "utils.hpp"
 
-size_t extract_rows_nb_csv(std::string path_p)
+int extract_rows_nb_csv(std::string path_p)
 {
-    size_t rows = 0;
+    int rows = 0;
     std::ifstream file(path_p);
     std::string line;
 
@@ -12,7 +12,7 @@ size_t extract_rows_nb_csv(std::string path_p)
     return rows;
 }
 
-Mission* extract_mission_csv(size_t size_p, std::string path_p)
+Mission* extract_mission_csv(int size_p, std::string path_p)
 {
     Mission* missions = new Mission[size_p];
     std::ifstream file(path_p);
@@ -21,7 +21,7 @@ Mission* extract_mission_csv(size_t size_p, std::string path_p)
     Skills skill;
 
 
-    for(size_t i=0; i<size_p; i++){
+    for(int i=0; i<size_p; i++){
         getline(file, s_id, ',');
         getline(file, s_day, ',');
         getline(file, s_starting_period, ',');
@@ -44,7 +44,7 @@ Mission* extract_mission_csv(size_t size_p, std::string path_p)
     return missions;
 }
 
-Employee* extract_employee_csv(size_t size_p, std::string path_p)
+Employee* extract_employee_csv(int size_p, std::string path_p)
 {
     Employee* employees = new Employee[size_p];
     std::ifstream file(path_p);
@@ -52,7 +52,7 @@ Employee* extract_employee_csv(size_t size_p, std::string path_p)
     Specialties specialty;
     Skills skill;
 
-    for(size_t i=0; i<size_p; i++){
+    for(int i=0; i<size_p; i++){
         getline(file, s_id, ',');
         getline(file, s_skill, ',');
         getline(file, s_specialty, ',');
@@ -73,14 +73,14 @@ Employee* extract_employee_csv(size_t size_p, std::string path_p)
     return employees;
 
 }
-float* extract_distance_matrix_csv(size_t size_p, std::string path_p)
+float* extract_distance_matrix_csv(int size_p, std::string path_p)
 {
     float* distances = new float[(size_p+1) * (size_p+1)];//+1 for sessad
     std::ifstream file(path_p);
     std::string data;
 
-    for(size_t i=0; i<size_p+1; i++){
-        for(size_t j=0; j< size_p; j++){
+    for(int i=0; i<size_p+1; i++){
+        for(int j=0; j< size_p; j++){
             getline(file, data, ',');
             distances[i*size_p + j] = std::stof(data);
         } 
@@ -92,10 +92,10 @@ float* extract_distance_matrix_csv(size_t size_p, std::string path_p)
  * print a matrix
  */
 template<typename T>
-void print_matrix(size_t size_p, const T* matrix_p)
+void print_matrix(int size_p, const T* matrix_p)
 {
-    for (size_t i=0; i < size_p; i++){
-        for (size_t j=0; j < size_p; j++){
+    for (int i=0; i < size_p; i++){
+        for (int j=0; j < size_p; j++){
             std::cout<<matrix_p[i*size_p + j] << " ";
         }
         std::endl;
